backend/requirements.txt
openai
requests
beautifulsoup4
fitz
psycopg2-binary
re